requirements.txt
google-adk
google-generativeai
gradio
pymupdf
pytesseract
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
reportlab
pillow
